+++ GreenhouseGraphing/src/models/associations/ReportTemplateDataPointAssoc.java
@@ -1,18 +1,11 @@
 package models.associations;
 
-/**
- * Created with IntelliJ IDEA.
- * User: matthallman
- * Date: 11/14/13
- * Time: 2:49 PM
- * To change this template use File | Settings | File Templates.
- */
 public class ReportTemplateDataPointAssoc {
 
-    long reportTemplateId;
-    long dataPointId;
+    Integer reportTemplateId;
+    Integer dataPointId;
 
-    public ReportTemplateDataPointAssoc(long reportTemplateId, long dataPointId) {
+    public ReportTemplateDataPointAssoc(Integer reportTemplateId, Integer dataPointId) {
         this.reportTemplateId = reportTemplateId;
         this.dataPointId = dataPointId;
     }
@@ -21,7 +14,7 @@ public class ReportTemplateDataPointAssoc {
         return reportTemplateId;
     }
 
-    public void setReportTemplateId(long reportTemplateId) {
+    public void setReportTemplateId(Integer reportTemplateId) {
         this.reportTemplateId = reportTemplateId;
     }
 
@@ -29,7 +22,7 @@ public class ReportTemplateDataPointAssoc {
         return dataPointId;
     }
 
-    public void setDataPointId(long dataPointId) {
+    public void setDataPointId(Integer dataPointId) {
         this.dataPointId = dataPointId;
     }
 }
